LICENSE.md: Add further terms & conditions
Our material has been prepared for informational and educational purposes
only, without regard to any particular user's investment objectives, financial
situation or means.  We are not soliciting any action based upon it. Our
material is not to be construed as a recommendation; or an offer to buy or
sell; or the solicitation of an offer to buy or sell any security, financial
product, or instrument.

You should neither construe any of our material as business, financial,
investment, hedging, trading, legal, regulatory, tax, or accounting advice nor
make our service the primary basis for any investment decisions made by or on
behalf of you, your accountants, or your managed or fiduciary accounts.