lib/ys_gauss_mix.py: fit volatilities w/ GM(2)
Add gm2_vols_fit(), gm2_vols(), and gm2_georet() to the module.

Fix gm2_strategy() when infeasible solution occurs.
A seemingly free parameter like b may not lead to a solution
in the real domain, but a slight upward adjustment may help
to avoid the imaginary domain, thanks to sympy.